% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_metric.R
\name{select_metric}
\alias{select_metric}
\title{select_metric for the server}
\usage{
select_metric(filtered_data, input)
}
\arguments{
\item{filtered_data}{filtered dataset from user input countries and date range}

\item{input}{reactivevalues object generated by shiny input functions in the UI part}
}
\description{
selects the metric for which values are to generated and outputted to the graph. eg - deaths, confirmed, tests, recovered
}
