% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_metric.R
\name{select_metric}
\alias{select_metric}
\title{select_metric}
\usage{
select_metric(selected_date, input)
}
\arguments{
\item{selected_date}{date param}

\item{input}{input param}
}
\description{
select the metric
}
